Add company top level menu entry, company and employee groups
issue8819
review286291002
