+++ CHANGELOG
@@ -1,3 +1,6 @@
+* Add company and employee administrator group
+* Add top level menu entry
+
 Version 5.4.0 - 2019-11-04
 * Bug fixes (see mercurial logs for details)
