--- cron.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0"?>
-<!-- This file is part of Tryton.  The COPYRIGHT file at the top level of
-this repository contains the full copyright notices and license terms. -->
-<tryton>
-    <data>
-        <record model="ir.ui.view" id="cron_view_form">
-            <field name="model">ir.cron</field>
-            <field name="inherit" ref="ir.cron_view_form"/>
-            <field name="name">cron_form</field>
-        </record>
-    </data>
-</tryton>
