+++ res.py
@@ -2,95 +2,10 @@
 # this repository contains the full copyright notices and license terms.
 import copy
 
-from trytond.model import ModelView, ModelSQL, fields, tree
-from trytond.wizard import Wizard, StateView, Button, StateTransition
-from trytond.report import Report
-from trytond.pyson import Eval, If
+from trytond.model import ModelSQL, fields
+from trytond.pool import PoolMeta, Pool
+from trytond.pyson import Eval
 from trytond.transaction import Transaction
-from trytond.pool import Pool, PoolMeta
-
-try:
-    import pytz
-    TIMEZONES = [(x, x) for x in pytz.common_timezones]
-except ImportError:
-    TIMEZONES = []
-TIMEZONES += [(None, '')]
-
-Transaction.cache_keys.update({'company', 'employee'})
-
-__all__ = ['Company', 'Employee', 'UserEmployee', 'User',
-    'Sequence', 'SequenceStrict', 'Date', 'CompanyConfigStart',
-    'CompanyConfig', 'CompanyReport', 'LetterReport', 'Rule']
-
-
-class Company(tree(), ModelSQL, ModelView):
-    'Company'
-    __name__ = 'company.company'
-    party = fields.Many2One('party.party', 'Party', required=True,
-            ondelete='CASCADE')
-    parent = fields.Many2One('company.company', 'Parent',
-        help="Add the company below the parent.")
-    childs = fields.One2Many('company.company', 'parent', 'Children',
-        help="Add children below the company.")
-    header = fields.Text(
-        'Header', help="The text to display on report headers.")
-    footer = fields.Text(
-        'Footer', help="The text to display on report footers.")
-    currency = fields.Many2One('currency.currency', 'Currency', required=True,
-        help="The main currency for the company.")
-    timezone = fields.Selection(TIMEZONES, 'Timezone', translate=False,
-        help="Used to compute the today date.")
-    employees = fields.One2Many('company.employee', 'company', 'Employees',
-        help="Add employees to the company.")
-
-    def get_rec_name(self, name):
-        return self.party.rec_name
-
-    @classmethod
-    def search_rec_name(cls, name, clause):
-        return [('party.rec_name',) + tuple(clause[1:])]
-
-    @classmethod
-    def write(cls, companies, values, *args):
-        super(Company, cls).write(companies, values, *args)
-        # Restart the cache on the domain_get method
-        Pool().get('ir.rule')._domain_get_cache.clear()
-
-
-class Employee(ModelSQL, ModelView):
-    'Employee'
-    __name__ = 'company.employee'
-    party = fields.Many2One('party.party', 'Party', required=True,
-        help="The party which represents the employee.")
-    company = fields.Many2One('company.company', 'Company', required=True,
-        help="The company to which the employee belongs.")
-    start_date = fields.Date('Start Date',
-        domain=[If((Eval('start_date')) & (Eval('end_date')),
-                    ('start_date', '<=', Eval('end_date')),
-                    (),
-                )
-            ],
-        depends=['end_date'],
-        help="When the employee joins the company.")
-    end_date = fields.Date('End Date',
-        domain=[If((Eval('start_date')) & (Eval('end_date')),
-                    ('end_date', '>=', Eval('start_date')),
-                    (),
-                )
-            ],
-        depends=['start_date'],
-        help="When the employee leaves the company.")
-
-    @staticmethod
-    def default_company():
-        return Transaction().context.get('company')
-
-    def get_rec_name(self, name):
-        return self.party.rec_name
-
-    @classmethod
-    def search_rec_name(cls, name, clause):
-        return [('party.rec_name',) + tuple(clause[1:])]
 
 
 class UserEmployee(ModelSQL):
@@ -284,104 +199,3 @@ class User(metaclass=PoolMeta):
         super(User, cls).write(*args)
         # Restart the cache on the domain_get method
         Rule._domain_get_cache.clear()
-
-
-class Sequence(metaclass=PoolMeta):
-    __name__ = 'ir.sequence'
-    company = fields.Many2One('company.company', 'Company',
-        domain=[
-            ('id', If(Eval('context', {}).contains('company'), '=', '!='),
-                Eval('context', {}).get('company', -1)),
-            ], help="Restrict the sequence usage to the company.")
-
-    @classmethod
-    def __setup__(cls):
-        super(Sequence, cls).__setup__()
-        cls._order.insert(0, ('company', 'ASC'))
-
-    @staticmethod
-    def default_company():
-        return Transaction().context.get('company')
-
-
-class SequenceStrict(Sequence):
-    __name__ = 'ir.sequence.strict'
-
-
-class Date(metaclass=PoolMeta):
-    __name__ = 'ir.date'
-
-    @classmethod
-    def today(cls, timezone=None):
-        pool = Pool()
-        Company = pool.get('company.company')
-        company_id = Transaction().context.get('company')
-        if timezone is None and company_id:
-            company = Company(company_id)
-            if company.timezone:
-                timezone = pytz.timezone(company.timezone)
-        return super(Date, cls).today(timezone=timezone)
-
-
-class CompanyConfigStart(ModelView):
-    'Company Config'
-    __name__ = 'company.company.config.start'
-
-
-class CompanyConfig(Wizard):
-    'Configure Company'
-    __name__ = 'company.company.config'
-    start = StateView('company.company.config.start',
-        'company.company_config_start_view_form', [
-            Button('Cancel', 'end', 'tryton-cancel'),
-            Button('OK', 'company', 'tryton-ok', True),
-            ])
-    company = StateView('company.company',
-        'company.company_view_form', [
-            Button('Cancel', 'end', 'tryton-cancel'),
-            Button('Add', 'add', 'tryton-ok', True),
-            ])
-    add = StateTransition()
-
-    def transition_add(self):
-        User = Pool().get('res.user')
-
-        self.company.save()
-        users = User.search([
-                ('main_company', '=', None),
-                ])
-        User.write(users, {
-                'main_company': self.company.id,
-                'company': self.company.id,
-                })
-        return 'end'
-
-
-class CompanyReport(Report):
-
-    @classmethod
-    def get_context(cls, records, data):
-        report_context = super(CompanyReport, cls).get_context(records, data)
-        report_context['company'] = report_context['user'].company
-        return report_context
-
-
-class LetterReport(CompanyReport):
-    __name__ = 'party.letter'
-
-    @classmethod
-    def execute(cls, ids, data):
-        with Transaction().set_context(address_with_party=True):
-            return super(LetterReport, cls).execute(ids, data)
-
-
-class Rule(metaclass=PoolMeta):
-    __name__ = 'ir.rule'
-
-    @classmethod
-    def _get_cache_key(cls):
-        key = super(Rule, cls)._get_cache_key()
-        # XXX Use company from context instead of browse to prevent infinite
-        # loop, but the cache is cleared when User is written.
-        return key + (Transaction().context.get('company'),
-            Transaction().context.get('employee'))
