+++ tryton.cfg
@@ -7,4 +7,4 @@ depends:
     res
 xml:
     company.xml
-    cron.xml
+    ir.xml
